Modularization  - Separated classes into different files.
